fix: extract Monaco vitest mock to satisfy ui-lint line limit
Follow-up to coverage mergeability commit; keeps mount-handler wiring
without exceeding noExcessiveLinesPerFunction in vitest.setup.ts.
